+++ menus/defaults.lua
@@ -14,6 +14,8 @@ local default_data = {
 	movement_smoothing = true,
 	teleport_on_release = false,
 	rotation_requires_press = false,
+	rotation_delay = 0.15,
+	rotation_amount = 45,
 
 	-- Camera fading parameters
 	cam_fade_distance = 0,